db-api/app/requirements.txt
fastapi[standard]
mysql-connector-python